+++ hw14-spring-mvc/src/main/java/ru/otus/vsh/hw14/webCore/server/Routes.java
@@ -2,10 +2,6 @@ package ru.otus.vsh.hw14.webCore.server;
 
 public final class Routes {
 
-    public static final String API_LOGIN = "/api/login/";
-    public static final String API_USER = "/api/user/";
-    public static final String API_ROLE = "/api/role/";
-
     public static final String PAGE_INDEX = "/";
     public static final String PAGE_ACTIONS = "/actions";
     public static final String PAGE_USERS = "/users";
